requirements.txt
django>=3.1
djangorestframework
pymysql
bcrypt
python-dotenv
djangorestframework-jwt
cryptography
PyJWT==2.1.0
sendgrid